Move changes to sync package
