+++ pkg/registry/apis/provisioning/jobs/sync/changes.go
@@ -1,4 +1,4 @@
-package jobs
+package sync
 
 import (
 	"fmt"
