+++ pkg/registry/apis/provisioning/jobs/sync/changes_test.go
@@ -1,4 +1,4 @@
-package jobs
+package sync
 
 import (
 	"encoding/json"
