+++ pkg/registry/apis/provisioning/jobs/syncer.go
@@ -20,6 +20,7 @@ import (
 	"github.com/grafana/grafana/pkg/apis/dashboard"
 	folders "github.com/grafana/grafana/pkg/apis/folder/v0alpha1"
 	provisioning "github.com/grafana/grafana/pkg/apis/provisioning/v0alpha1"
+	"github.com/grafana/grafana/pkg/registry/apis/provisioning/jobs/sync"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/repository"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/resources"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/safepath"
@@ -174,7 +175,7 @@ func (r *syncJob) run(ctx context.Context) error {
 	if err != nil {
 		return fmt.Errorf("error reading tree: %w", err)
 	}
-	changes, err := Changes(source, target)
+	changes, err := sync.Changes(source, target)
 	if err != nil {
 		return fmt.Errorf("error calculating changes: %w", err)
 	}
@@ -195,7 +196,7 @@ func (r *syncJob) run(ctx context.Context) error {
 	return r.applyChanges(ctx, changes)
 }
 
-func (r *syncJob) applyChanges(ctx context.Context, changes []ResourceFileChange) error {
+func (r *syncJob) applyChanges(ctx context.Context, changes []sync.ResourceFileChange) error {
 	// Do the longest paths first (important for delete)
 	sort.Slice(changes, func(i, j int) bool {
 		return len(changes[i].Path) > len(changes[j].Path)
